compiled on windows

+++ scourge/src/scourge.cpp
@@ -928,7 +928,7 @@ bool Scourge::handleEvent(SDL_Event *event) {
 	  setPlayer(3);
     }
     else if(ea == SET_PLAYER_ONLY && 
-            battleTurn < battleRound.size() &&
+            battleTurn < (int)battleRound.size() &&
             !session->getUserConfiguration()->isBattleTurnBased()) { 
         party->togglePlayerOnly();
     }    
@@ -2494,7 +2494,7 @@ bool Scourge::handlePartyEvent(Widget *widget, SDL_Event *event) {
   } else if(widget == player4Button) {
     setPlayer(Constants::PLAYER_4 - Constants::PLAYER_1);
   } else if(widget == groupButton && 
-            battleTurn < battleRound.size() &&
+            battleTurn < (int)battleRound.size() &&
             !session->getUserConfiguration()->isBattleTurnBased()) {
     party->togglePlayerOnly();
   } else if(widget == roundButton) {
